Generalize support for SQLTransform
